added link validation process and page redirect to login page after success register
